+++ src/main/java/prestudy/framework/spring/api/config/WebConfig.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import prestudy.framework.spring.api.jwt.JwtInterceptor;
+import prestudy.framework.spring.api.authenticate.jwt.JwtInterceptor;
 
 @Configuration
 @RequiredArgsConstructor
