+++ src/test/java/prestudy/framework/spring/api/authenticate/jwt/JwtProviderTest.java
@@ -1,10 +1,11 @@
-package prestudy.framework.spring.api.jwt;
+package prestudy.framework.spring.api.authenticate.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import prestudy.framework.spring.api.authenticate.jwt.JwtProvider;
 import prestudy.framework.spring.support.IntegrationTestSupport;
 
 import java.time.LocalDateTime;
