+++ src/test/java/prestudy/framework/spring/api/service/comment/CommentServiceTest.java
@@ -45,7 +45,7 @@ class CommentServiceTest extends IntegrationTestSupport {
     @Test
     void createCommentWithInvalidToken() {
         // given
-        given(jwtRequestUtils.getUserId()).willThrow(new IllegalStateException("토큰이 유효하지 않습니다."));
+        given(authenticationUserProvider.getUserId()).willThrow(new IllegalStateException("토큰이 유효하지 않습니다."));
 
         CommentCreateCommand command = CommentCreateCommand.builder()
             .boardId(1L)
@@ -62,7 +62,7 @@ class CommentServiceTest extends IntegrationTestSupport {
     @Test
     void createCommentWithInvalidUser() {
         // given
-        given(jwtRequestUtils.getUserId()).willReturn(0L);
+        given(authenticationUserProvider.getUserId()).willReturn(0L);
 
         CommentCreateCommand command = CommentCreateCommand.builder()
             .boardId(1L)
@@ -83,7 +83,7 @@ class CommentServiceTest extends IntegrationTestSupport {
 
         userRepository.save(user);
 
-        given(jwtRequestUtils.getUserId()).willReturn(user.getId());
+        given(authenticationUserProvider.getUserId()).willReturn(user.getId());
 
         CommentCreateCommand command = CommentCreateCommand.builder()
             .boardId(1L)
@@ -106,7 +106,7 @@ class CommentServiceTest extends IntegrationTestSupport {
         Board board = createBoardEntity(user);
         boardRepository.save(board);
 
-        given(jwtRequestUtils.getUserId()).willReturn(user.getId());
+        given(authenticationUserProvider.getUserId()).willReturn(user.getId());
 
         CommentCreateCommand command = CommentCreateCommand.builder()
             .boardId(board.getId())
@@ -124,7 +124,7 @@ class CommentServiceTest extends IntegrationTestSupport {
     @Test
     void updateCommentWithInvalidToken() {
         // given
-        given(jwtRequestUtils.getUserId()).willThrow(new IllegalStateException("토큰이 유효하지 않습니다."));
+        given(authenticationUserProvider.getUserId()).willThrow(new IllegalStateException("토큰이 유효하지 않습니다."));
 
         CommentUpdateCommand command = CommentUpdateCommand.builder()
             .id(1L)
@@ -141,7 +141,7 @@ class CommentServiceTest extends IntegrationTestSupport {
     @Test
     void updateCommentWithInvalidUser() {
         // given
-        given(jwtRequestUtils.getUserId()).willReturn(0L);
+        given(authenticationUserProvider.getUserId()).willReturn(0L);
 
         CommentUpdateCommand command = CommentUpdateCommand.builder()
             .id(1L)
@@ -161,7 +161,7 @@ class CommentServiceTest extends IntegrationTestSupport {
         User user = User.ofUser("123abc", "Password12!");
         userRepository.save(user);
 
-        given(jwtRequestUtils.getUserId()).willReturn(user.getId());
+        given(authenticationUserProvider.getUserId()).willReturn(user.getId());
 
         CommentUpdateCommand command = CommentUpdateCommand.builder()
             .id(0L)
@@ -186,7 +186,7 @@ class CommentServiceTest extends IntegrationTestSupport {
         Board board = createBoardEntity(commentUser);
         boardRepository.save(board);
 
-        given(jwtRequestUtils.getUserId()).willReturn(tokenUser.getId());
+        given(authenticationUserProvider.getUserId()).willReturn(tokenUser.getId());
 
         Comment comment = Comment.builder()
             .content("댓글 생성")
@@ -219,7 +219,7 @@ class CommentServiceTest extends IntegrationTestSupport {
         Board board = createBoardEntity(user);
         boardRepository.save(board);
 
-        given(jwtRequestUtils.getUserId()).willReturn(admin.getId());
+        given(authenticationUserProvider.getUserId()).willReturn(admin.getId());
 
         Comment comment = Comment.builder()
             .content("댓글 생성")
@@ -251,7 +251,7 @@ class CommentServiceTest extends IntegrationTestSupport {
         Board board = createBoardEntity(user);
         boardRepository.save(board);
 
-        given(jwtRequestUtils.getUserId()).willReturn(user.getId());
+        given(authenticationUserProvider.getUserId()).willReturn(user.getId());
 
         Comment comment = Comment.builder()
             .content("댓글 생성")
@@ -277,7 +277,7 @@ class CommentServiceTest extends IntegrationTestSupport {
     @Test
     void deleteCommentWithInvalidToken() {
         // given
-        given(jwtRequestUtils.getUserId()).willThrow(new IllegalStateException("토큰이 유효하지 않습니다."));
+        given(authenticationUserProvider.getUserId()).willThrow(new IllegalStateException("토큰이 유효하지 않습니다."));
 
         CommentDeleteCommand command = CommentDeleteCommand.of(1L);
 
@@ -291,7 +291,7 @@ class CommentServiceTest extends IntegrationTestSupport {
     @Test
     void deleteCommentWithInvalidUser() {
         // given
-        given(jwtRequestUtils.getUserId()).willReturn(0L);
+        given(authenticationUserProvider.getUserId()).willReturn(0L);
 
         CommentDeleteCommand command = CommentDeleteCommand.of(1L);
 
@@ -309,7 +309,7 @@ class CommentServiceTest extends IntegrationTestSupport {
 
         userRepository.save(user);
 
-        given(jwtRequestUtils.getUserId()).willReturn(user.getId());
+        given(authenticationUserProvider.getUserId()).willReturn(user.getId());
 
         CommentDeleteCommand command = CommentDeleteCommand.of(1L);
 
@@ -331,7 +331,7 @@ class CommentServiceTest extends IntegrationTestSupport {
         Board board = createBoardEntity(commentUser);
         boardRepository.save(board);
 
-        given(jwtRequestUtils.getUserId()).willReturn(tokenUser.getId());
+        given(authenticationUserProvider.getUserId()).willReturn(tokenUser.getId());
 
         Comment comment = Comment.builder()
             .content("댓글 생성")
@@ -362,7 +362,7 @@ class CommentServiceTest extends IntegrationTestSupport {
 
         boardRepository.save(board);
 
-        given(jwtRequestUtils.getUserId()).willReturn(admin.getId());
+        given(authenticationUserProvider.getUserId()).willReturn(admin.getId());
 
         Comment comment = Comment.builder()
             .content("댓글 생성")
@@ -392,7 +392,7 @@ class CommentServiceTest extends IntegrationTestSupport {
         Board board = createBoardEntity(user);
         boardRepository.save(board);
 
-        given(jwtRequestUtils.getUserId()).willReturn(user.getId());
+        given(authenticationUserProvider.getUserId()).willReturn(user.getId());
 
         Comment comment = Comment.builder()
             .content("댓글 생성")
