+++ src/test/java/prestudy/framework/spring/support/IntegrationTestSupport.java
@@ -3,12 +3,12 @@ package prestudy.framework.spring.support;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
-import prestudy.framework.spring.api.jwt.JwtRequestUtils;
+import prestudy.framework.spring.api.authenticate.AuthenticationUserProvider;
 
 @ActiveProfiles("test")
 @SpringBootTest
 public abstract class IntegrationTestSupport {
 
     @MockitoBean
-    protected JwtRequestUtils jwtRequestUtils;
+    protected AuthenticationUserProvider authenticationUserProvider;
 }
